alphaedge/requirements.txt
aio-pika==9.4.2
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aioredis==2.0.1
aiormq==6.8.0
aiosignal==1.3.1
anyio==4.4.0
arrow==1.3.0
async-timeout==4.0.3
attrs==24.2.0
bidict==0.23.1
binaryornot==0.4.4
breeze_connect==1.0.56
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==1.7.3
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.2.2
fastapi==0.112.0
frozenlist==1.4.1
h11==0.14.0
httptools==0.6.1
idna==3.6
Jinja2==3.1.2
jinja2-time==0.2.0
loguru==0.7.2
manage-fastapi==1.1.1
MarkupSafe==2.1.3
multidict==6.0.5
numpy==2.0.1
pamqp==3.3.0
pandas==2.2.2
pika==1.3.2
poyo==0.5.0
prompt-toolkit==3.0.41
pydantic==1.10.13
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.9.1
python-slugify==8.0.1
python-socketio==5.11.3
pytz==2024.1
PyYAML==6.0.2
questionary==1.10.0
redis==5.0.8
requests==2.31.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
text-unidecode==1.3
typer==0.4.2
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2024.1
urllib3==2.1.0
uvicorn==0.30.5
uvloop==0.20.0
watchfiles==0.23.0
wcwidth==0.2.12
websocket-client==1.8.0
websockets==12.0
wsproto==1.2.0
yarl==1.9.4
